pyproject.toml
[project]
name = "adhd-linguistic-patterns-beliefs-portuguese-women"
version = "0.1.0"
description = "Master Thesis for Computer Science and Engineering at Instituto Superior Técnico"
readme = "README.md"
requires-python = "==3.10.12"
dependencies = [
    "ace-tools>=0.0",
    "baycomp>=1.0.3",
    "bertopic[flair,gensim,vision,visualization]>=0.17.3",
    "bokeh>=3.7.2",
    "ctransformers[cuda]>=0.2.27",
    "deep-translator>=1.11.4",
    "dotenv>=0.9.9",
    "faker>=37.4.0",
    "jinja2>=3.1.6",
    "liwc>=0.5.0",
    "lxml>=5.4.0",
    "matplotlib>=3.10.1",
    "mistral-common>=1.8.3",
    "numpy>=2.2.5",
    "openai==1.99.9",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pingouin>=0.5.5",
    "pip>=25.1",
    "pre-commit>=4.2.0",
    "presidio-analyzer>=2.2.358",
    "presidio-anonymizer>=2.2.358",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.6.1",
    "scikit-posthocs>=0.11.4",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "shap>=0.48.0",
    "spacy>=3.8.5",
    "stanza>=1.10.1",
    "statsmodels>=0.14.4",
    "tensor>=0.3.6",
    "textblob>=0.19.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "umap-learn>=0.5.9.post2",
    "wordcloud>=1.9.4",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "ruff>=0.9.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
